EventManager implements ActionListener
Added an ActionListener called EventManager to the Board class. The
PlayBoard currently registers its Swipe buttons to the EventManager,
but the only functionality of the Buttons is to output a message to the
console confirming the pressing of that button. Also, a variety of
cosmetic changed to documentation, as well as a small bit of clarity of
documentation has been added.